requirements.txt
yt-dlp
google-genai
python-dotenv>=1.0.1
pdf2image>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.1
